Perform compare on CHERI architectures.
Update the compare of a capability register
and an immediate value. On CHERI architectures
the capability value can only be compared against
an immediate value, by first getting the value
(with a gcvalue). The x1 register is listed in
the asm block clobbers list, and as it isn't
used for any other purspose in the remiander of
the block it is used to hold the capability
value for the purpose of the compare.